Moves AttributionNetwork
