+++ RevenueCat/Scripts/AttributionNetwork.cs
@@ -1,0 +1,16 @@
+using System.Diagnostics.CodeAnalysis;
+
+public partial class Purchases
+{
+    [SuppressMessage("ReSharper", "InconsistentNaming")]
+    [SuppressMessage("ReSharper", "UnusedMember.Global")]
+    public enum AttributionNetwork
+    {
+        APPLE_SEARCH_ADS = 0,
+        ADJUST = 1,
+        APPSFLYER = 2,
+        BRANCH = 3,
+        TENJIN = 4,
+        FACEBOOK = 5
+    }
+}
